requirements.txt
coreapi==2.1.1
Django==1.10.3
django-cors-headers==1.3.1
django-extra-fields==0.8
django-filter==1.0.1
django-rest-swagger==2.1.0
djangorestframework==3.5.3
djangorestframework-bulk==0.2.1
djangorestframework-jwt==1.8.0
drfdocs==0.0.11
httplib2==0.9.2
itypes==1.1.0
mysqlclient==1.3.9
openapi-codec==1.2.0
Pillow==3.4.2
PyJWT==1.4.2
PySocks==1.6.5
pytz==2016.10
requests==2.12.3
simplejson==3.10.0
six==1.10.0
sms==0.1
twilio==5.6.0
uritemplate==3.0.0
